Refactoring and removed DataSourceFactory
