--- src/main/java/org/example/lib/DBTypes.java
@@ -1,6 +1,0 @@
-package org.example.lib;
-
-public enum DBTypes {
-    POSTGRE,
-    H2;
-}
