+++ src/main/resources/db.properties
@@ -1,4 +1,3 @@
 jdbc-url=jdbc:h2:./ORMAnager
 jdbc-username=sa
 jdbc-password=password
-jdbc-typeDB=h2
