--- src/main/resources/h2mem.properties
@@ -1,4 +1,0 @@
-jdbc-url=jdbc:h2:file:./ormdb
-jdbc-username=
-jdbc-password=
-jdbc-typeDB=H2
