+++ src/test/java/org/example/model/AnnotationsUtilsTest.java
@@ -1,0 +1,32 @@
+package org.example.model;
+
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+class AnnotationsUtilsTest {
+
+    @Test
+    void isAnnotationPresent() {
+    }
+
+    @Test
+    void testIsAnnotationPresent() {
+    }
+
+    @Test
+    void getNameOfColumn() {
+    }
+
+    @Test
+    void getIdType() {
+    }
+
+    @Test
+    void getNameOfTable() {
+    }
+
+    @Test
+    void getEnumType() {
+    }
+}
